chore(odyssey-react): fix type annotations for form
